retrieve and insert embedding

+++ priv/migrations/20251121143525-embeddings.sql
@@ -3,7 +3,7 @@ CREATE EXTENSION IF NOT EXISTS vector;
 
 CREATE TABLE item_embeddings
 (
-    item_id    INTEGER PRIMARY KEY,
+    item_id    BIGINT PRIMARY KEY,
     embedding  vector(2048) NOT NULL,
     model      TEXT         NOT NULL,
     created_at TIMESTAMPTZ  NOT NULL DEFAULT NOW(),
